analyse_test/requirements.txt
# Core LangGraph dependencies
langgraph>=0.2.0
langchain-core>=0.1.0
langchain-aws>=0.1.0

# AWS SDK
boto3>=1.34.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Fix protobuf compatibility
protobuf>=5.29.3,<6.0.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
